reinstalled node modules for validator dependency that was not in auth me, added validations and corrected constraints for models and migrations, created seed files and seeded only demoUser

+++ backend/db/migrations/20240624175833-create-venue.js
@@ -38,11 +38,11 @@ module.exports = {
 					allowNull: false,
 				},
 				lat: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				lng: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				createdAt: {
